added processedeventpurger
