+++ src/BoltOn/Cqrs/CqrsInterceptor.cs
@@ -2,12 +2,9 @@
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using BoltOn.Data;
 using BoltOn.Logging;
 using BoltOn.Mediator.Interceptors;
 using BoltOn.Mediator.Pipeline;
-using BoltOn.UoW;
-using System.Transactions;
 
 namespace BoltOn.Cqrs
 {
@@ -16,19 +13,16 @@ namespace BoltOn.Cqrs
 		private readonly EventBag _eventBag;
 		private readonly IBoltOnLogger<CqrsInterceptor> _logger;
 		private readonly IEventDispatcher _eventDispatcher;
-		private readonly ICqrsRepositoryFactory _cqrsRepositoryFactory;
-		private readonly IUnitOfWorkManager _unitOfWorkManager;
+		private readonly IProcessedEventPurger _processedEventPurger;
 
 		public CqrsInterceptor(EventBag eventBag, IBoltOnLogger<CqrsInterceptor> logger,
-			IEventDispatcher eventDispatcher, 
-			ICqrsRepositoryFactory cqrsRepositoryFactory,
-			IUnitOfWorkManager unitOfWorkManager)
+			IEventDispatcher eventDispatcher,
+			IProcessedEventPurger processedEventPurger)
 		{
 			_eventBag = eventBag;
 			_logger = logger;
 			_eventDispatcher = eventDispatcher;
-			_cqrsRepositoryFactory = cqrsRepositoryFactory;
-			_unitOfWorkManager = unitOfWorkManager;
+			_processedEventPurger = processedEventPurger;
 		}
 
 		public TResponse Run<TRequest, TResponse>(IRequest<TResponse> request,
@@ -57,27 +51,7 @@ namespace BoltOn.Cqrs
 
 			if (request is CqrsEvent cqrsEvent)
 			{
-				_logger.Debug($"Building repository. SourceType: {cqrsEvent.SourceTypeName}");
-                var method = _cqrsRepositoryFactory.GetType().GetMethod("GetRepository");
-				var sourceEntityType = Type.GetType(cqrsEvent.SourceTypeName);
-				var generic = method.MakeGenericMethod(sourceEntityType);
-				dynamic repository = generic.Invoke(_cqrsRepositoryFactory, null);
-				_logger.Debug("Built repository");
-				using (var uow = _unitOfWorkManager.Get(u => u.TransactionScopeOption = TransactionScopeOption.RequiresNew))
-				{
-					var cqrsEntity = await repository.GetByIdAsync(cqrsEvent.SourceId);
-					var baseCqrsEntity = cqrsEntity as BaseCqrsEntity;
-					_logger.Debug($"Fetched BaseCqrsEntity. Id: {cqrsEvent.SourceId}");
-					var @event = baseCqrsEntity?.Events.FirstOrDefault(f => f.Id == cqrsEvent.Id);
-					if (@event != null)
-					{
-						_logger.Debug("Removing event...");
-						baseCqrsEntity.Events.Remove(@event);
-						await repository.UpdateAsync(cqrsEntity);
-						_logger.Debug($"Removed event. Id: {@event.Id}");
-					}
-					uow.Commit();
-				}
+				await _processedEventPurger.PurgeAsync(cqrsEvent);
 			}
 
 			return response;
